+++ StorageNetwork/ModApi.cs
@@ -1,7 +1,0 @@
-﻿namespace StorageNetwork {
-    public class ModApi : IModApi {
-        public void InitMod(Mod _modInstance) {
-            throw new System.NotImplementedException();
-        }
-    }
-}